refactor to support react 15.6
