--- node_modules_temp/@types/react/package.json
@@ -1,46 +1,0 @@
-{
-    "name": "@types/react",
-    "version": "15.0.20",
-    "description": "TypeScript definitions for React",
-    "license": "MIT",
-    "contributors": [
-        {
-            "name": "Asana",
-            "url": "https://asana.com"
-        },
-        {
-            "name": "AssureSign",
-            "url": "http://www.assuresign.com"
-        },
-        {
-            "name": "Microsoft",
-            "url": "https://microsoft.com"
-        },
-        {
-            "name": "John Reilly",
-            "url": "https://github.com/johnnyreilly/"
-        },
-        {
-            "name": "Benoit Benezech",
-            "url": "https://github.com/bbenezech"
-        },
-        {
-            "name": "Patricio Zavolinsky",
-            "url": "https://github.com/pzavolinsky"
-        },
-        {
-            "name": "Digiguru",
-            "url": "https://github.com/digiguru"
-        }
-    ],
-    "main": "",
-    "repository": {
-        "type": "git",
-        "url": "https://www.github.com/DefinitelyTyped/DefinitelyTyped.git"
-    },
-    "scripts": {},
-    "dependencies": {},
-    "peerDependencies": {},
-    "typesPublisherContentHash": "4025623f969a6998c810b8cb38bed4b3d2ab06c6b285b9129e7ee59c81d93fe5",
-    "typeScriptVersion": "2.1"
-}
