+++ src/demo/advanced/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import Workspace from './Workspace';
 import AppBar from './AppBar';
 import Drafting from './Drafting';
@@ -6,16 +6,19 @@ import Settings from './Settings';
 import Keybinding from './Keybinding';
 import Canvas from './Canvas';
 
-import { Provider } from '../../lib'
+import { Provider } from '../../lib';
 
-export const Advanced = () =>
+export const Advanced = () => (
   <Provider>
-    <Workspace />
-    <AppBar />
-    <Drafting />
-    <Settings />
-    <Keybinding />
-    <Canvas />
+    <div>
+      <Workspace />
+      <AppBar />
+      <Drafting />
+      <Settings />
+      <Keybinding />
+      <Canvas />
+    </div>
   </Provider>
+);
 
 export default Advanced;
